Add unit test for algo error issue
