+++ bundles/table-editor/pom.xml
@@ -126,5 +126,17 @@
             <artifactId>spatial-utilities</artifactId>
             <version>${h2-gis-version}</version>
         </dependency>
+        <dependency>
+            <groupId>junit</groupId>
+            <artifactId>junit</artifactId>
+            <version>4.10</version>
+            <scope>test</scope>
+        </dependency>
+        <dependency>
+            <groupId>org.orbisgis</groupId>
+            <artifactId>h2spatial</artifactId>
+            <version>${h2-gis-version}</version>
+            <scope>test</scope>
+        </dependency>
     </dependencies>
 </project>
